chore: ignore engine start if runtime is already up
Thanks to the engine overhaul, we shut down the runtime static at the
end of every session. We should check and make sure that's the case
before letting the user start the engine again. Ideally the UX would
handle this with not allowing the button to be active but eh.