chore: Update device configs to store in rust
If we keep our user device configs stored in Rust,
we can use the native heap to act as shared memory
between dart isolates. Holding a
Arc<DeviceConfigurationManager> on the heap as a
lazy_static allows ut to update it (via DashMap
interior mutability) while Engines are running,
possibly in an isolate.